// Hawa-front/src/app/shared/model/history-forest-specailOrCommune.model.ts
import { TreeSpecies } from './dictionary/tree-species.model';
import { Dictionary } from './dictionary/dictionary.model';
import { Actor } from './dictionary/actor.model';
import { ActorType } from './dictionary/actor-type.model';
import { UserBuyerHistoryModel } from './user-buyer/user-buyer-history.model';

export class HistoryForestSpecailOrCommune {
    id: number;
    treeSpec: TreeSpecies;
    compartment: Dictionary;
    compartmentCode: String;
    subCompartment: Dictionary;
    subCompartmentCode: string;
    plot: Dictionary;
    plotCode: String;
    volumnPerPlot: number;
    area: number;
    plantingYear: number;
    dispute: Dictionary;
    actor: Actor;
    actorType: ActorType;
    forestCert: Dictionary;
    reliability: Dictionary;
    landUseCert: Dictionary;
    conflictSitCode: number;
    locationLatitude: number;
    locationLongitude: number;
    plantingDate: number;
    user: UserBuyerHistoryModel;
    createDate: number;
}